Bulk
- Rename 'src' to 'predicting_glucose_measurement'
- Removed setup py (redundant, since poetry)
- Add main as entrypoint in poetry
- Update packages
