+++ predicting_glucose_levels/__init__.py
@@ -1,0 +1,3 @@
+from predicting_glucose_levels.bootstrap import bootstrap_di
+
+bootstrap_di()
