+++ predicting_glucose_levels/data/ingestion/ingester.py
@@ -3,9 +3,11 @@ from typing import List
 
 from kink import inject
 
-from src.data.ingestion.loader.abstract_loader import AbstractLoader
-from src.data.storage.abstract_storage import AbstractStorage
-from src.data.table_metadata import TableMetadata
+from predicting_glucose_levels.data.ingestion.loader.abstract_loader import (
+    AbstractLoader,
+)
+from predicting_glucose_levels.data.storage.abstract_storage import AbstractStorage
+from predicting_glucose_levels.data.table_metadata import TableMetadata
 
 
 @inject
