+++ pyproject.toml
@@ -5,7 +5,7 @@ description = ""
 authors = ["Aucke Bos <aucke.bos97@gmail.com>"]
 license = "MIT"
 packages = [
-    { include = "src", from="." },
+    { include = "predicting_glucose_levels", from="." },
 ]
 
 [tool.poetry.dependencies]
@@ -27,9 +27,9 @@ kink = "^0.6.6"
 autoflake = "^2.2.0"
 
 
-[tool.poetry.dev-dependencies]
-
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
+[tool.poetry.scripts]
+glucose = "predicting_glucose_levels.main:cli"
