--- setup.py
@@ -1,10 +1,0 @@
-from setuptools import find_packages, setup
-
-setup(
-    name="predicting-glucose-levels",
-    packages=find_packages(),
-    version="0.1.0",
-    description="A short description of the project.",
-    author="AuckeBos",
-    license="MIT",
-)
