--- src/__init__.py
@@ -1,3 +1,0 @@
-from src.bootstrap import bootstrap_di
-
-bootstrap_di()
